Another 1s compilation speedup.
